+++ tests/report_text.rs
@@ -21,6 +21,7 @@ pub const REPO: &str = "repo-4e3ae97";
 /// Command under test against one repo provided by env
 pub fn cargo_bin(repo: &str) -> assert_cmd::Command {
     let mut c = common::cargo_bin();
+    c.env_clear();
     c.envs(DEFAULT_ENV);
     c.env("BORG_REPO", repo);
     c
@@ -56,6 +57,8 @@ fn no_backups() {
 fn check_ok() {
     init::one_archive(REPO, "{utcnow}Z");
     cargo_bin(REPO)
+        .env("BORGREPORT_CHECK", "true")
+        .env("BORGREPORT_COMPACT", "false")
         .assert()
         .stdout(predicate::str::contains(REPO).count(2))
         .stdout(predicate::str::contains("borg check"))
@@ -99,4 +102,36 @@ fn archive_too_old() {
         .stdout(predicate::str::contains(format!("{REPO}: Last backup is older than")).count(1))
         .stdout(predicate::str::contains("borg check"))
         .stdout(predicate::str::contains("yes |").count(1));
+}
+
+/// A repository with two valid archives using globs compacts OK once.
+#[sealed_test]
+fn compact_ok() {
+    init::two_archives(REPO, "etc-{utcnow}Z", "srv-{utcnow}Z");
+    cargo_bin(REPO)
+        .env("BORGREPORT_GLOB_ARCHIVES", "etc-* srv-*")
+        .env("BORGREPORT_CHECK", "false")
+        .env("BORGREPORT_COMPACT", "true")
+        .assert()
+        .stdout(predicate::str::contains("etc-").count(1))
+        .stdout(predicate::str::contains("srv-").count(1))
+        .stdout(predicate::str::contains(REPO).count(3))
+        .stdout(predicate::str::contains("borg compact").count(1));
+}
+
+/// A repository with a faulty archive compacts NOT OK when check has failed.
+#[sealed_test]
+fn compact_not_ok() {
+    init::faulty_archive(REPO, "{utcnow}Z");
+    cargo_bin(REPO)
+        .env("BORGREPORT_CHECK", "true")
+        .env("BORGREPORT_COMPACT", "true")
+        .assert()
+        .stdout(predicate::str::contains(REPO).count(4))
+        .stdout(predicate::str::contains("borg compact").count(1))
+        .stdout(
+            predicate::str::contains("Repositories with errors or warnings are not compacted.")
+                .count(1),
+        )
+        .stdout(predicate::str::contains("- |").count(2));
 }