+++ CHANGELOG.md
@@ -8,7 +8,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/bbx0/borgreport/compare/v0.3.0...HEAD) - 2025-03-08
+## [Unreleased](https://github.com/bbx0/borgreport/compare/v0.3.0...HEAD) - 2025-06-02
+
+### Added
+
+- Option `--compact` to perform `borg compact` on a repository.
 
 ### Changed